requirements.txt
ocrmypdf
pandas
pdfminer.six
boto3
tqdm
pyPDF2
python-dotenv